pyproject.toml
[project]
name = "SDVValentines2025"
version = "0.1.0"
description = "Discord bot for the SDV server's 2025 Valentine's event"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.4.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
